requirements.txt
aiogram
python-dotenv
aiohttp
discord.py
yeelight
requests
praw